requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
transformers
torch
numpy
soundfile
librosa
gtts
pypdf
gradio_client
python-multipart
faiss-cpu
sentence-transformers
